fix -Werrors and more, favour size_t to int
+ that's all the Werrors except unused params and missing initializers.
+ switching from signed to unsigned is probably gonna bite me at some
  point, but meh.
+ also use size_t in stb_sb.h, not that i was ever gonna make one >2gb.
+ changed stb_sb.h growth factor to 1.5 for no good reason.
+ fixed 1byte stack overflow in mod_quotes which clang's asan caught.
+ fixed !vod double print if the control chars are equal
+ fixed the dlopen error printing logic so it actually prints the
  dlerror msg if one exists.

+++ src/inso_ht.h
@@ -285,7 +285,7 @@ inso_htpriv_del_i(inso_ht* ht, intptr_t idx){
 
 		INSO_HT_DBG("hash=%zu, idx2=%zu, ", hash, idx2);
 
-		if(idx2 <= idx){
+		if(idx2 <= (size_t)idx){
 			INSO_HT_DBG("swapping.");
 			memcpy(
 				ht->memory + idx  * ht->elem_size,
